311: Create guaranteed QoS and set JAVA_OPTS (#21)
* 311: Create guaranteed QoS and set JAVA_OPTS

Huge amount of CPU being requested at start up caused node issue and
eviction of pods. We need to change the way tomcat is started such
that it respects the container memory limits and has a guaranteed
quality of service.

Set the following java opts;
-XX:+UseG1GC -XX:+UseContainerSupport -XX:MaxRAMPercentage=50

Update the chart version as extension/v1beta1 is old and I was getting
warnings that is deprecated, and it simply won't deploy on newer
versions of k8s, e.g. in local minikube.

Note: These changes dramatically increase the start up time of the
config server.This meant the pods were starting to fail as the liveness probe
wasn't responding in time. To address that I have increased the
intitalDelaySeconds value on the liveness probe

Issue: SecureBankingAccessToolkit/SecureBankingAccessToolkit#311

* sdlkfj